Fix for the GPU again

+++ keras/layers/recurrent.py
@@ -55,7 +55,7 @@ class SimpleRNN(Layer):
         mask = T.neq(X, self.mask_val).sum(axis=2) > 0 # (time, nb_samples) matrix with a 1 for every unmasked entry
         mask = T.addbroadcast(mask[:, :, np.newaxis], 2)
 
-        mask_tm1 = alloc_zeros_matrix(*mask.shape)
+        mask_tm1 = alloc_zeros_matrix(*mask.shape).astype('int8')
         mask_tm1 = T.addbroadcast(T.set_subtensor(mask_tm1[1:, :, :], mask[:-1, :, :]), 2)
 
         x = T.dot(X, self.W) + self.b
